refactor : Image 컴포넌트를 common 폴더로 분리리
