+++ src/components/header/registerIcon.js
@@ -1,13 +1,16 @@
 import { $ } from "../../utils/dom";
+import Image from "../common/image";
 
 const registerIcon = () => {
   const registerIcon = document.createElement("button");
-  registerIcon.classList.add("gnb__button");
 
-  registerIcon.innerHTML = `<img src="./add-button.png" alt="음식점 추가" />`;
+  registerIcon.classList.add("gnb__button");
+  registerIcon.appendChild(Image("./add-button.png", "음식점 추가"));
+
   registerIcon.addEventListener("click", () => {
     $(".modal-backdrop").classList.add("open");
   });
+
   return registerIcon;
 };
